requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
apscheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
streamlit==1.28.1
pydantic==2.5.0
websockets 